newest

+++ src/main/java/com/example/projetcir3/Position.java
@@ -1,9 +1,9 @@
 package com.example.projetcir3;
 
 public class Position {
-    public int x;
-    public int y;
-    Position(int x, int y) {
+    public float x;
+    public float y;
+    Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
